work in progress

+++ lib/timer-text-widget.dart
@@ -1,9 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:endorphin/running-page.dart';
 import 'dart:async';
 
 class TimerText extends StatefulWidget {
-  TimerText({Key key}) : super(key: key);
+  final DateTime startTime;
+  final bool active;
+
+  TimerText(this.startTime, this.active, {Key key}) : super(key: key);
 
   @override
   _TimerTextState createState() => new _TimerTextState();
@@ -11,55 +13,46 @@ class TimerText extends StatefulWidget {
 
 class _TimerTextState extends State<TimerText> {
   Timer timer;
-  String hours = '0';
-  String minutes = '00';
-  String seconds = '00';
+  String hours, minutes, seconds;
 
-  _TimerTextState() {
-    timer = new Timer.periodic(new Duration(milliseconds: 1000), callback);
+  @override
+  void initState() {
+    timer = Timer.periodic(Duration(milliseconds: 1000), timerCallback);
+    _countTime();
+    super.initState();
   }
 
-  void callback(Timer timer) {
-    if (stopwatch.isRunning) {
-      setState(() {
-        int milliseconds = stopwatch.elapsedMilliseconds;
-        hours = (milliseconds / 3600000).truncate().toString();
-        minutes = ((milliseconds / 60000).truncate() % 60).toString().padLeft(2, '0');
-        seconds = ((milliseconds / 1000).truncate() % 60).toString().padLeft(2, '0');
-      });
+  @override
+  void dispose() {
+    timer.cancel();
+    timer = null;
+    super.dispose();
+  }
+
+  void timerCallback(Timer timer) {
+    _countTime();
+    setState(() {});
+  }
+
+  _countTime() {
+    if(widget.active) {
+      Duration runDuration = DateTime.now().difference(widget.startTime);
+      int milliseconds = runDuration.inMilliseconds;
+      hours = runDuration.inHours.toString();
+      minutes =
+          ((milliseconds / 60000).truncate() % 60).toString().padLeft(2, '0');
+      seconds =
+          ((milliseconds / 1000).truncate() % 60).toString().padLeft(2, '0');
     }
   }
 
   @override
   Widget build(BuildContext context) {
-    TextStyle hoursMinutesTextStyle =
-        new TextStyle(fontSize: 90.0, color: statColor);
-    TextStyle secondsTextStyle =
-        new TextStyle(fontSize: 50.0, color: statColor, height: 1.9);
-    return new Row(children: [
-      new Text(hours + ":" + minutes, style: hoursMinutesTextStyle),
-      new Text(":" + seconds, style: secondsTextStyle)
+    TextStyle hoursMinutesTextStyle = TextStyle(fontSize: 90.0);
+    TextStyle secondsTextStyle = TextStyle(fontSize: 50.0, height: 1.9);
+    return Row(children: [
+      Text(hours + ":" + minutes, style: hoursMinutesTextStyle),
+      Text(":" + seconds, style: secondsTextStyle)
     ]);
-  }
-}
-
-class TimerTextFormatter {
-  static String getHoursMinutes(int milliseconds) {
-    int hundreds = (milliseconds / 10).truncate();
-    int seconds = (hundreds / 100).truncate();
-    int minutes = (seconds / 60).truncate();
-    int hours = (minutes / 60).truncate();
-
-    String hoursStr = (hours % 60).toString();
-    String minutesStr = (minutes % 60).toString().padLeft(2, '0');
-
-    return "$hoursStr:$minutesStr";
-  }
-
-  static String getSeconds(int milliseconds) {
-    int hundreds = (milliseconds / 10).truncate();
-    int seconds = (hundreds / 100).truncate();
-    String secondsStr = (seconds % 60).toString().padLeft(2, '0');
-    return ":$secondsStr";
   }
 }
